feat(upload): add image preview delete
On deleting an image preview, it will also delete the associated file
that was uploaded onto Supabase

- Implement image preview delete functionality in `uploadForm.js`
- Add route for deleting supabase files in `fileRouter.js`
- Add controller to deal with the actual deleting logic in
  `fileController.js`
- Remove forgotten console.log() in `utils.js`

NOTE: After some mulling over, I realized that with my current tools
and the limitations of this project's structure, implementing an image
preview like the ones seen on social media sites is currently out of my
scope. It's also quite complex for a quick project meant for learning
Prisma ORM and not file handling like this one. Therefore, I will
recognize this sunk-cost fallacy and save it for a future portfolio
piece when I can implement React and other tools to make it possible.
Therefore I will stop here and cut my scope in order to finish on time.
I hope you understand!